feat(wrapper)!: simplify prediction using the pyfunc wrapper
